document_translator/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
azure-core==1.30.2
azure-identity==1.16.1
azure-servicebus==7.12.2
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
Django==4.2.13
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
idna==3.7
isodate==0.6.1
msal==1.28.1
msal-extensions==1.1.0
multidict==6.0.5
packaging==24.1
portalocker==2.8.2
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing-extensions==4.12.2
tzdata==2024.1
whitenoise==6.7.0
yarl==1.9.4
django-environ==0.11.2
pylint==3.2.3